+++ src/main/java/live/databo3/front/dto/GetNotificationDto.java
@@ -1,4 +1,4 @@
-package live.databo3.front.admin.dto;
+package live.databo3.front.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class GetNotificationListResponse {
+public class GetNotificationDto {
     private Long notificationId;
     private String title;
-    private String author;
+    private String contents;
     private LocalDateTime date;
-
+    private String memberId;
 }
